+++ src/Koffee/Utility.fs
@@ -9,6 +9,9 @@ open System.Reactive.Concurrency
 open FSharp.Control
 open Acadian.FSharp
 
+let inline fstf f (a, b) = (f a, b)
+let inline sndf f (a, b) = (a, f b)
+
 let runAsync (f: unit -> 'a) = async {
     let ctx = System.Threading.SynchronizationContext.Current
     do! Async.SwitchToNewThread()
@@ -117,4 +120,3 @@ module Rect =
           Top = top
           Width = width
           Height = height }
-